Some what working AlphaBeta implementation with option to choose which player is AI. Fixed too many bugs to even count. AI IS WORK IN PROGRESS!

+++ trunk/Chess/Position.h
@@ -120,6 +120,6 @@ public:
 	bool bIsCheck(UI64 BitBoards[]);
 
 	//The board evaluation functions
-	double evaluate(UI64 BitBoards[]);
+	int evaluate(UI64 BitBoards[]);
 };
